% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateFreqNA.R
\name{EstimateFreqNA}
\alias{EstimateFreqNA}
\title{EstimateFreqNA}
\usage{
EstimateFreqNA(path.mcmc)
}
\arguments{
\item{path.mcmc}{path to directory containing stored MCMC computations}
}
\value{
A vector of length \code{nloc} (the number of loci) whose entries  are estimated frequencies of null alleles.
}
\description{
Estimate frequencies of null alleles at each locus.
}
