% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geneland-package.R
\docType{package}
\name{Geneland-package}
\alias{Geneland-package}
\alias{Geneland}
\title{Geneland: clustering of georeferenced genetic and morphometric data}
\description{
Geneland implements Bayesian models and Markov chain Monte Carlo (MCMC)
algorithms to infer population genetic structure from georeferenced
individual multi-locus genotypes and quantitative phenotypes.
}
\section{Getting started}{

[MCMC()] runs the sampler and writes its output to a directory,
[PostProcessChain()] post-processes a chain to deal with label switching,
and the `Plot*` functions and [PosteriorMode()] display the results. See
`vignette("Geneland")` for a worked example.
}

\section{Models for allele frequencies}{

The `freq.model` argument of [MCMC()] selects how allele frequencies are
treated across populations.

\describe{
  \item{`"Uncorrelated"`}{Frequencies in each population are given
    independent Dirichlet priors. This is the safer default: it does not
    assume the populations share a common history, and it is the only
    model compatible with `filter.null.alleles = TRUE`.}
  \item{`"Correlated"`}{Frequencies are drawn around those of a common
    ancestral population, each population having its own drift parameter
    with a `Beta(shape1, shape2)` prior. This model has more power to
    detect subtle structure, but it is also more prone to detecting
    spurious clusters when its assumptions are violated.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gilles-guillot/Geneland}
  \item Report bugs at \url{https://github.com/gilles-guillot/Geneland/issues}
}

}
\author{
\strong{Maintainer}: Gilles Guillot \email{gilles.b.guillot@gmail.com} (\href{https://orcid.org/0000-0003-0675-8325}{ORCID})

Authors:
\itemize{
  \item Gilles Guillot \email{gilles.b.guillot@gmail.com} (\href{https://orcid.org/0000-0003-0675-8325}{ORCID})
}

}
\keyword{internal}
