% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullify.R
\name{nullify}
\alias{nullify}
\title{nullify}
\usage{
nullify(genotypes, nall.null = 1, nloc.null)
}
\arguments{
\item{genotypes}{a matrix of genotypes as produced by simFmodel and simIBD}

\item{nall.null}{number of null alleles on each locus}

\item{nloc.null}{number of loci carrying null alleles}
}
\value{
A list with one component, `genotypes`: the new genotypes after
  alteration.
}
\description{
Simulates null alleles
}
