\name{PlanarProj}
\alias{PlanarProj}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Planar projection of set of points}
\description{Project a set of points whose coordinates are given as
  longitudes and latitudes onto a planar domain}
\usage{
PlanarProj(lon,lat)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{lon}{A vector of longitude in decimal degrees}
  \item{lat}{A vector of latitude in decimal degrees}
}
\value{A list with elements x and y}



\references{
  \url{http://en.wikipedia.org/wiki/Mercator_projection#Mathematics_of_the_projection}
  }

\author{Gilles Guillot  }

\keyword{}
