% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDrift.R
\name{PlotDrift}
\alias{PlotDrift}
\title{PlotDrift}
\usage{
PlotDrift(path.mcmc, printit = FALSE, file)
}
\arguments{
\item{path.mcmc}{Path to output files directory}

\item{printit}{Logical : if TRUE, figures are also printed}

\item{file}{Character : Path to file where figures
should be printed}
}
\description{
Gives a plot of the trace of the drift factors along the
  MCMC run
}
