% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFreqA.R
\name{PlotFreqA}
\alias{PlotFreqA}
\title{PlotFreqA}
\usage{
PlotFreqA(path.mcmc, iloc, iall, printit = FALSE, path)
}
\arguments{
\item{path.mcmc}{Path to output files directory}

\item{iloc}{Integer number : index of locus}

\item{iall}{Integer number : index of allele}

\item{printit}{Logical : if TRUE, figures are also printed}

\item{path}{Character : Path to directory where figures
should be printed}
}
\description{
Plot frequency of  allele \code{iall}  of locus \code{iloc}
  in ancestral population
}
