% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotntile.R
\name{Plotntile}
\alias{Plotntile}
\title{Plotntile}
\usage{
Plotntile(path.mcmc, burnin, printit = FALSE, file)
}
\arguments{
\item{path.mcmc}{Path to output files directory}

\item{burnin}{An integer: number of saved iterations to discard for
the representation of the histogram of the chain}

\item{printit}{Logical : if TRUE, figures are also printed}

\item{file}{Character : Path to file where figures
should be printed}
}
\description{
Plot of number of tiles in the Poisson-Voronoi tessellation
  along the MCMC run
}
