% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTessDyn.R
\name{PlotTessDyn}
\alias{PlotTessDyn}
\title{PlotTessDyn}
\usage{
PlotTessDyn(coordinates = NULL, path.mcmc, nxgrid, nygrid)
}
\arguments{
\item{coordinates}{coordinates}

\item{path.mcmc}{Path to output files directory}

\item{nxgrid}{nb of pixels in the x direction}

\item{nygrid}{nb of pixels in the y direction}
}
\description{
Plot sequence of tessellations simulated along an MCMC run.
}
