\name{rdiscr}
\alias{rdiscr}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Internal function for the simulation of a categorical random variable}
\description{Simulation one value from categorical distribution}
\usage{
rdiscr(f)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{f}{Vector of frequencies (sum(\code{f}) should be equal to one}
}
\details{

}
\value{
 An integer drawn from the categorical distribution
}


\keyword{}