% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFmodel.R
\name{simFmodel}
\alias{simFmodel}
\title{Simulation of multi-locus genetic data from the spatial F-model}
\usage{
simFmodel(
  nindiv,
  coordinates,
  coord.lim = c(0, 1, 0, 1),
  number.nuclei,
  coord.nuclei,
  color.nuclei,
  nall,
  npop,
  freq.model = "Uncorrelated",
  drift,
  dominance = "Codominant",
  plots = FALSE,
  ploth = FALSE
)
}
\arguments{
\item{nindiv}{Simulation of multi-locus genetic data from the spatial F-model}

\item{coordinates}{Matrix (2 rows, nindiv columns) of spatial coordinates of individuals}

\item{coord.lim}{Vector of limits of spatial domain to be considered (x min, x max, y min, y max)}

\item{number.nuclei}{Integer: number of nuclei in the Voronoi tessellation}

\item{coord.nuclei}{Coordinates of nuclei of Voronoi tessellation}

\item{color.nuclei}{Population labels of the nuclei (vector of integers of size number.nuclei)}

\item{nall}{Vector of integers giving number of alleles at each locus}

\item{npop}{Number of populations}

\item{freq.model}{model for frequencies:"Correlated" or "Uncorrelated"}

\item{drift}{Vector (of size \code{npop}) of drift factors between 0 and 1 (only for the Correlated model)}

\item{dominance}{A character string "Codominant" or "Dominant". If
"Dominant" is chosen, the first allele is treated as a recessive
allele and all heterozigous are converted into homozigous for the
dominant allele. The presence of the "dominant" allele is coded as 1,
the absence of the "dominant" allele is coded as 0.}

\item{plots}{Logical: if TRUE, spatial coordinates are ploted}

\item{ploth}{Logical: if TRUE, barplots for allele frequencies are ploted}
}
\value{
A list of variables involved in the simulation. The elements of
this list are: 
  coordinates, genotypes, allele.numbers, number.nuclei, coord.nuclei, color.nuclei, 
frequencies,  ancestral.frequencies, drifts, index.nearest.nucleus
}
\description{
Simulation of multi-locus genetic data from the spatial F-model
}
