% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatGenotypes.R
\name{FormatGenotypes}
\alias{FormatGenotypes}
\title{FormatGenotypes}
\usage{
FormatGenotypes(genotypes, ploidy)
}
\value{
A list with elements: 
    genotypes (matrix with one line per individual and
      one or two columns per locus with alleles coded by integers) and 
 allele.numbers (a vector giving the number of possible alleles per locus
 @export
}
\description{
Takes genotype data as a matrix with
  one line per individual and two columns per locus,
  with alleles coded by integers (number of replications for
                                  micro-satellites data).
  Build a new matrix with alleles codes as consecutive integers.
  If a locus  has 7 alleles they will be coded as 1,2,...7.
  Since version 1.0.1, this function does not have to be called by
  users. It is called through MCMC.
 @param genotypes A matrix with
    one line per individual and two columns per locus,
    with alleles coded by integers
 @param ploidy 1 or 2
}
