\name{Geneland.GUI}
\alias{Geneland.GUI}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Graphical interface for package Geneland }
\description{
 Launch a menu driven interface to package Geneland
}
\usage{
Geneland.GUI(lib.loc)}
\arguments{
  \item{lib.loc}{A character string giving the path to the directory
  where Geneland is installed}
}
%- maybe also 'usage' for other objects documented here.
\author{Filipe Santos}

