% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geneland.GUI.R
\name{Geneland.GUI}
\alias{Geneland.GUI}
\title{Geneland.GUI}
\usage{
Geneland.GUI(lib.loc = NULL)
}
\arguments{
\item{lib.loc}{A character string giving the path to the directory
where Geneland is installed}
}
\description{
Launch a menu driven interface to package Geneland
}
