% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl2gp.R
\name{gl2gp}
\alias{gl2gp}
\title{gl2gp}
\usage{
gl2gp(coordinates, genotypes, file)
}
\value{
An ascci file is written. This file may need to be
  process in order to convert eol character from unix style to dos style. 
  @export
}
\description{
Takes the matrices  of  coordinates and genotypes in the Geneland format
  and writes it as an ascci file in the Genepop format
  \@param coordinates   Matrix of coordinates
  \@param genotypes  MAtrix of genotypes
  \@param file   Character string giving the path to the file
    where the data in Genepop format should be written
}
