% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullify.R
\name{nullify}
\alias{nullify}
\title{nullify}
\usage{
nullify(genotypes, nall.null = 1, nloc.null)
 @param genotypes  a matrix of genotypes as produced by simFmodel and simIBD
 @param nall.null number of null alleles on each locus
 @param nloc.null number of loci carrying null alleles
}
\value{
A list with component: 
   @param genotypes  The new genotypes after alteration.
    @export
}
\description{
Simulates null alleles
}
