\name{PlotDrift}
\alias{PlotDrift}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Plot the drift factors}
\description{Gives a plot of the trace of the drift factors along the
  MCMC run}
\usage{
PlotDrift(repmcmc,printit=FALSE,file=NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{repmcmc}{ Path to output files directory }
   \item{printit}{Logical : if TRUE, figures are also printed}
  \item{file}{Character : Path to file where figures
    should be printed}
}

\references{A spatial statistical model for landscape genetics, Guillot,
  Estoup, Mortier, Cosson, Genetics, 2005

 Guillot, G., Geneland : A program for landscape genetics.
 Molecular Ecology  Notes, submited.
}
\author{Gilles Guillot  }


\keyword{}