\name{PlotDrift}
\alias{PlotDrift}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Plot the drift factors}
\description{Gives a plot of the trace of the drift factors along the
  MCMC run}
\usage{
PlotDrift(path.mcmc,printit=FALSE,file)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{path.mcmc}{ Path to output files directory }
   \item{printit}{Logical : if TRUE, figures are also printed}
  \item{file}{Character : Path to file where figures
    should be printed}
}


\author{Gilles Guillot  }

