% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFreqA.R
\name{PlotFreqA}
\alias{PlotFreqA}
\title{PlotFreqA}
\usage{
PlotFreqA(path.mcmc,iloc,iall,printit=FALSE,path)
 @param path.mcmc  Path to output files directory 
 @param iloc Integer number : index of locus
 @param iall Integer number : index of allele
 @param printit Logical : if TRUE, figures are also printed
 @param path Character : Path to directory where figures
    should be printed
}
\description{
Plot frequency of  allele \code{iall}  of locus \code{iloc}
  in ancestral population
}
