\name{Plotnpop}
\alias{Plotnpop}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Plot of number of populations along the MCMC run}
\description{Gives a plot of the number of populations along the MCMC run
}
\usage{
Plotnpop(path.mcmc,printit=FALSE,file=NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{path.mcmc}{ Path to output files directory }
  \item{printit}{Logical : if TRUE, figures are also printed}
  \item{file}{Character : Path to file where figures
    should be printed}
}

\references{A spatial statistical model for landscape genetics, Guillot,
Estoup, Mortier, Cosson, Genetics, 2005
  
 Guillot, G., Geneland : A program for landscape genetics.
 Molecular Ecology  Notes, submited.}
\author{Gilles Guillot  }

\keyword{}
