% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotnpop.R
\name{Plotnpop}
\alias{Plotnpop}
\title{Plotnpop}
\usage{
Plotnpop(path.mcmc,burnin,printit=FALSE,file,format="pdf")
  @param path.mcmc  Path to output files directory 
  @param printit Logical : if TRUE, figures are also printed
  @param file Character : Path to file where figures
    should be printed
  @param format format of the output file, should be either \code{"ps"}
    or \code{"pdf"}
  @param burnin An integer: number of saved iterations to discard for
    the representation of the histogram of the chain
}
\description{
Gives a plot of the number of populations along the MCMC run
}
