% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotntile.R
\name{Plotntile}
\alias{Plotntile}
\title{Plotntile}
\usage{
Plotntile(path.mcmc,burnin,printit=FALSE,file)
 @param path.mcmc  Path to output files directory 
 @param printit Logical : if TRUE, figures are also printed
 @param file Character : Path to file where figures
    should be printed
 @param burnin An integer: number of saved iterations to discard for
    the representation of the histogram of the chain
 @export
}
\description{
Plot of number of tiles in the Poisson-Voronoi tessellation
  along the MCMC run
}
