\name{PlotTessDyn}
\alias{PlotTessDyn}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Plot tessellation along an MCMC run.}

\description{Plot sequence of tessellations simulated along an MCMC run.}

\usage{
PlotTessDyn(coordinates,path.mcmc,nxgrid,nygrid)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{coordinates}{coordinates}
  \item{path.mcmc}{Path to output files directory }
  \item{nxgrid}{nb of pixels in the x direction}
  \item{nygrid}{nb of pixels in the y direction}
}


\author{Gilles Guillot}




