% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTessDyn.R
\name{PlotTessDyn}
\alias{PlotTessDyn}
\title{PlotTessDyn}
\usage{
PlotTessDyn(coordinates = NULL, path.mcmc, nxgrid, nygrid)
}
\description{
Plot sequence of tessellations simulated along an MCMC run.
  @param coordinates coordinates
  @param path.mcmc Path to output files directory 
  @param nxgrid nb of pixels in the x direction
  @param nygrid nb of pixels in the y direction
}
