% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostProcessChain.R
\name{PostProcessChain}
\alias{PostProcessChain}
\title{PostProcessChain}
\usage{
PostProcessChain(coordinates,
                   path.mcmc,nxdom, nydom,burnin)
  @param coordinates Spatial coordinates of individuals. A matrix with 2
    columns and one line per individual.
  
  @param path.mcmc  Path to output files directory 
  
  @param nxdom Number of pixel for discretization of the spatial domain
    in the horizontal direction
  
  @param nydom Number of pixel for discretization of the spatial domain
    in the vertical direction
  
  @param burnin Number of iterations of the chain to throw away.
    WARNING : this argument should be given the number of stored
    iterations (and not the number of computed iterations which differ
                if \code{thinning} !=1). If you have
    \code{nit}=100000 and \code{thinning}=100, then only 1000 iterations
    are stored. Then \code{burnin}=10 will throw away 10 stored
    iterations, namely 100*10 computed iterations.
}
\description{
Computes posterior probabilities of population membership for each pixel of the spatial domain.
}
\details{
Posterior probability of population membership for each pixel: 
    They  are
    written in an ascii file called \file{proba.pop.membership.txt}.
    Two first columns are coordinates of pixels then 
    one column per population (thus \code{npopmax} values are computed
                               for each pixel). 
    Images in each column of \file{proba.pop.membership.txt} are stored
    starting from the bottom left pixel.
    First line of \file{proba.pop.membership.txt}  = bottom left pixel ,
    second line of \file{proba.pop.membership.txt} = upward neighboor of
    the previous pixel, etc...
  
  
  Posterior probability of population membership for each
    individual:  They are written in a file named
      \file{proba.pop.membership.indiv.txt}. One line per individuals and
      \code{2+npopmax} columns per individual. Two first columns are
      spatial coordinates.
    
  
  Label of modal population for pixels and individuals: 
    They are written in files named \file{modal.pop.txt} and
    \file{modal.pop.indiv.txt}    respectively.
}
