% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.simdata.R
\name{show.simdata}
\alias{show.simdata}
\title{@title show.simdata
 @description Graphical display of data simulated by simdata
  @param dataset An R object produced by function \code{simdata}
  @param plot.coord Logical
  @param file.plot.coord Character string. Path to the file where the
    graphic should be stored
  @param plot.tess  Logical
  @param file.plot.tess Character string. Path to the file where the
    graphic should be stored
  @param plot.freq.grid Logical
  @param file.plot.freq.grid Character string. Path to the file where the
    graphic should be stored
  @param loc.grid Vector of integers giving indices of loci for which the
    map of frequencies is required
  @param plot.freq.indiv Logical
  @param file.plot.freq.indiv Character string. Path to the file where the
    graphic should be stored
  @param loc.indiv Vector of integers giving indices of loci for which the
    plot of frequencies at sites of individuals is required
  @param plot.gen Logical
  @param file.plot.gen Character string. Path to the file where the
    graphic should be stored
  @param zlim.freq A vector of two integers giving the limit of the
    values for the image of fields of frequencies}
\usage{
show.simdata(dataset,
plot.coord = FALSE,
file.plot.coord,
plot.tess = FALSE,
file.plot.tess,
plot.freq.grid = FALSE,
file.plot.freq.grid,
loc.grid = 1,
plot.freq.indiv = FALSE,
file.plot.freq.indiv,
loc.indiv=1,
zlim.freq=c(0,1),
plot.gen = FALSE,
file.plot.gen)
}
\description{
@title show.simdata
 @description Graphical display of data simulated by simdata
  @param dataset An R object produced by function \code{simdata}
  @param plot.coord Logical
  @param file.plot.coord Character string. Path to the file where the
    graphic should be stored
  @param plot.tess  Logical
  @param file.plot.tess Character string. Path to the file where the
    graphic should be stored
  @param plot.freq.grid Logical
  @param file.plot.freq.grid Character string. Path to the file where the
    graphic should be stored
  @param loc.grid Vector of integers giving indices of loci for which the
    map of frequencies is required
  @param plot.freq.indiv Logical
  @param file.plot.freq.indiv Character string. Path to the file where the
    graphic should be stored
  @param loc.indiv Vector of integers giving indices of loci for which the
    plot of frequencies at sites of individuals is required
  @param plot.gen Logical
  @param file.plot.gen Character string. Path to the file where the
    graphic should be stored
  @param zlim.freq A vector of two integers giving the limit of the
    values for the image of fields of frequencies
}
